move types to types.go
